Add files via upload
This C program reverses an array of integers entered by the user. It prompts the user to input a specified number of elements, stores them in an array, and then reverses the order of the elements before displaying the reversed array.

Features
Accepts a user-defined number of integer elements.
Reverses the order of the elements in the array.
Displays the reversed array to the user.
How It Works
The program starts by displaying a title message.
It prompts the user to enter the number of elements they wish to input.
It initializes an array of the specified size and prompts the user to enter each element.
The reversearray function is called to reverse the elements of the array:
It uses two pointers, start and end, to swap elements from the beginning and end of the array until they meet in the middle.
After reversing, the program prints the reversed array.